requirements.txt
#fastapi==0.109.1
requests==2.28.1
google-auth==2.15.0
google-cloud-bigquery==3.25.0
uvicorn==0.22.0
pytest==8.3.3
#pydantic==1.10.2
pydantic[dotenv]==1.10.2
numpy==1.24.4
tenacity==8.2.3
